backend/agents/intake/requirements.txt
anthropic
python-dotenv
rich